more fixes, reorganizing before the big one
