+++ src/brain_input/vision/descriptors.rs
