+++ tests/tests_vision.rs
@@ -2,7 +2,7 @@ use feagi_core_data_structures_and_processing::brain_input::vision::single_frame
 use std::fs::File;
 use ndarray::Array3;
 use ndarray_npy::NpzReader;
-use feagi_core_data_structures_and_processing::brain_input::vision::single_frame_processing::*;
+use feagi_core_data_structures_and_processing::brain_input::vision::descriptors::*;
 
 #[test]
 fn test_loading_data_frame() {
